change solartime to module for Sun class...
because of the way that variables need to be passed back and forth between Sun object and SolarTime methods, SolarTime could not be a subclass of Sun. Almost as stupid to make them separate classes. Decided to split SolarTime methods in to module because
they are clearly a separate family, but are closely connected to the sun. Could just leave them in sun.rb, but for organisational reasons, put them in separate module