remove all trace of serialize...
could not see the benefit over just saving as a string and manipulating it after it is retrieved. Turning it in to an array upon retrieval is a no-brainer. Saving it as an array or json confuses me. Saving as a hash has benefits, but manipulating the csv file to contain hashes sounds boring